Convert key binding-related commands to new command API
