+++ src/engine/client/cl_keys.cpp
@@ -44,8 +44,6 @@ key up events are sent even if in console mode
 
 using Keyboard::Key;
 
-#define CLIP(t) Math::Clamp( (t), 0, MAX_TEAMS - 1 )
-
 bool key_overstrikeMode;
 bool bindingsModified;
 
@@ -60,14 +58,6 @@ static struct {
 	bool     valid;
 	int          check;
 } plusCommand;
-
-
-#define FIELD_TEAM            1
-#define FIELD_TEAM_SPECTATORS 2
-#define FIELD_TEAM_DEFAULT    4
-
-static void Field_CompleteKeyname( int flags );
-static void Field_CompleteTeamname( int flags );
 
 
 /*
@@ -424,437 +414,6 @@ bool Key_IsDown( Key key )
 	return keys[ key ].down;
 }
 
-
-
-/*
-===================
-Key_GetTeam
-Assumes 'three' teams: spectators, aliens, humans
-===================
-*/
-static const char *const teamName[] = { "default", "aliens", "humans", "others" };
-
-int Key_GetTeam( const char *arg, const char *cmd )
-{
-	static const struct {
-		char team;
-		char label[11];
-	} labels[] = {
-		{ 0, "spectators" },
-		{ 0, "default" },
-		{ 1, "aliens" },
-		{ 2, "humans" }
-	};
-	int t, l;
-
-	if ( !*arg ) // empty string
-	{
-		goto fail;
-	}
-
-	for ( t = 0; arg[ t ]; ++t )
-	{
-		if ( !Str::cisdigit( arg[ t ] ) )
-		{
-			break;
-		}
-	}
-
-	if ( !arg[ t ] )
-	{
-		t = atoi( arg );
-
-		if ( t != CLIP( t ) )
-		{
-			Log::Notice( "^3%s:^7 %d is not a valid team number\n", cmd, t );
-			return -1;
-		}
-
-		return t;
-	}
-
-	l = strlen( arg );
-
-	for ( unsigned t = 0; t < ARRAY_LEN( labels ); ++t )
-	{
-		// matching initial substring
-		if ( !Q_strnicmp( arg, labels[ t ].label, l ) )
-		{
-			return labels[ t ].team;
-		}
-	}
-
-fail:
-	Log::Notice( "^3%s:^7 '%s^7' is not a valid team name\n", cmd, arg );
-	return -1;
-}
-
-/*
-===================
-Key_SetBinding
-
-team == -1 clears all bindings for the key, then sets the spec/global binding
-===================
-*/
-void Key_SetBinding( Key key, int team, const char *binding )
-{
-	char *lcbinding; // fretn - make a copy of our binding lowercase
-	// so name toggle scripts work again: bind x name BzZIfretn?
-	// resulted into bzzifretn?
-
-	if ( !key.IsBindable() )
-	{
-		return;
-	}
-
-	// free old bindings
-	if ( team == -1 )
-	{
-		// just the team-specific ones here
-		for ( team = MAX_TEAMS - 1; team; --team )
-		{
-			if ( keys[ key ].binding[ team ] )
-			{
-				Z_Free( keys[ key ].binding[ team ] );
-				keys[ key ].binding[ team ] = nullptr;
-			}
-		}
-		// team == 0...
-	}
-
-	team = CLIP( team );
-
-	if ( keys[ key ].binding[ team ] )
-	{
-		Z_Free( keys[ key ].binding[ team ] );
-	}
-
-	// set the new binding, if not null/empty
-	if ( binding && binding[ 0 ] )
-	{
-		// allocate memory for new binding
-		keys[ key ].binding[ team ] = CopyString( binding );
-		lcbinding = CopyString( binding );
-		Q_strlwr( lcbinding );  // saves doing it on all the generateHashValues in Key_GetBindingByString
-		Z_Free( lcbinding );
-	}
-	else
-	{
-		keys[ key ].binding[ team ] = nullptr;
-	}
-
-	bindingsModified = true;
-}
-
-/*
-===================
-Key_GetBinding
-
--ve team no. = don't return the default binding
-===================
-*/
-const char *Key_GetBinding( Key key, int team )
-{
-	const char *bind;
-
-	if ( !key.IsBindable() )
-	{
-		return nullptr;
-	}
-
-	if ( team <= 0 )
-	{
-		return keys[ key ].binding[ CLIP( -team ) ];
-	}
-
-	bind = keys[ key ].binding[ CLIP( team ) ];
-	return bind ? bind : keys[ key ].binding[ 0 ];
-}
-
-/*
-===================
-Key_Unbind_f
-===================
-*/
-static void Key_Unbind_f()
-{
-	int b = Cmd_Argc();
-	int team = -1;
-
-	if ( b < 2 || b > 3 )
-	{
-		Cmd_PrintUsage("[<team>] <key>", "remove commands from a key");
-		return;
-	}
-
-	if ( b > 2 )
-	{
-		team = Key_GetTeam( Cmd_Argv( 1 ), "unbind" );
-
-		if ( team < 0 )
-		{
-			return;
-		}
-	}
-
-	Key key = Key_StringToKeynum( Cmd_Argv( b - 1 ) );
-
-	if ( !key.IsBindable() )
-	{
-		Log::Notice( "\"%s\" isn't a valid key\n", Cmd_Argv( 1 ) );
-		return;
-	}
-
-	Key_SetBinding( key, team, nullptr );
-}
-
-/*
-===================
-Key_Unbindall_f
-===================
-*/
-static void Key_Unbindall_f()
-{
-	for (auto& kv : keys)
-	{
-		Key_SetBinding( kv.first, -1, nullptr );
-	}
-}
-
-/*
-===================
-Key_Bind_f
-===================
-*/
-static void Key_Bind_f()
-{
-	int        c;
-	const char *key;
-	const char *cmd = nullptr;
-	int        team = -1;
-
-	int teambind = !Q_stricmp( Cmd_Argv( 0 ), "teambind" );
-
-	c = Cmd_Argc();
-
-	if ( c < 2 + teambind )
-	{
-		Cmd_PrintUsage( teambind ? "<team> <key> [<command>]" : "<key> [<command>]", "attach a command to a key");
-		return;
-	}
-	else if ( c > 2 + teambind )
-	{
-		cmd = Cmd_Argv( 2 + teambind );
-	}
-
-	if ( teambind )
-	{
-		team = Key_GetTeam( Cmd_Argv( 1 ), teambind ? "teambind" : "bind" );
-
-		if ( team < 0 )
-		{
-			return;
-		}
-	}
-
-	key = Cmd_Argv( 1 + teambind );
-	Key b = Key_StringToKeynum( key );
-
-	if ( !b.IsBindable() )
-	{
-		Log::Notice( "\"%s\" isn't a valid key\n", key );
-		return;
-	}
-
-	key = Key_KeynumToString( b );
-
-	if ( !cmd )
-	{
-		if ( teambind )
-		{
-			if ( keys[ b ].binding[ team ] )
-			{
-				Log::Notice( "\"%s\"[%s] = %s\n", key, teamName[ team ], Cmd_QuoteString( keys[ b ].binding[ team ] ) );
-			}
-			else
-			{
-				Log::Notice( "\"%s\"[%s] is not bound\n", key, teamName[ team ] );
-			}
-		}
-		else
-		{
-			bool bound = false;
-			int      i;
-
-			for ( i = 0; i < MAX_TEAMS; ++i )
-			{
-				if ( keys[ b ].binding[ i ] )
-				{
-					Log::Notice( "\"%s\"[%s] = %s\n", key, teamName[ i ], Cmd_QuoteString( keys[ b ].binding[ i ] ) );
-					bound = true;
-				}
-			}
-
-			if ( !bound )
-			{
-				Log::Notice( "\"%s\" is not bound\n", key );
-			}
-		}
-
-
-		return;
-	}
-
-
-	if ( c <= 3 + teambind )
-	{
-		Key_SetBinding( b, team, cmd );
-	}
-	else
-	{
-		// set to 3rd arg onwards (4th if team binding)
-		Key_SetBinding( b, team, Cmd_ArgsFrom( 2 + teambind ) );
-	}
-}
-
-/*
-===================
-Key_EditBind_f
-===================
-*/
-static void Key_EditBind_f()
-{
-	std::u32string buf;
-	int            b;
-	int            team = -1;
-
-	b = Cmd_Argc();
-
-	if ( b < 2 || b > 3 )
-	{
-		Cmd_PrintUsage("[<team>] <key>", nullptr);
-		return;
-	}
-
-	if ( b > 2 )
-	{
-		team = Key_GetTeam( Cmd_Argv( 1 ), "editbind" );
-
-		if ( team < 0 )
-		{
-			return;
-		}
-	}
-
-	const char *key = Cmd_Argv( b - 1 );
-	Key k = Key_StringToKeynum( key );
-
-	if ( !k.IsBindable() )
-	{
-		Log::Notice( "\"%s\" isn't a valid key\n", key );
-		return;
-	}
-
-	if ( team >= 0 )
-	{
-		buf = Str::UTF8To32("/teambind ");
-		buf += Str::UTF8To32( teamName[ team ] );
-		buf += Str::UTF8To32(" ");
-	}
-	else
-	{
-		buf = Str::UTF8To32("/bind ");
-	}
-
-	buf += Str::UTF8To32( Key_KeynumToString( k ) );
-	buf += Str::UTF8To32(" ");
-
-	const char *binding = Key_GetBinding( k, -team );
-	if ( binding )
-	{
-		buf += Str::UTF8To32( Cmd::Escape( std::string( binding ) ) );
-	}
-
-	// FIXME: use text console if that's where the editbind command was entered
-	Con_OpenConsole_f();
-	g_consoleField.SetText( buf );
-}
-
-/*
-============
-Key_WriteBindings
-
-Writes lines containing "bind key value"
-============
-*/
-void Key_WriteBindings( fileHandle_t f )
-{
-	int team;
-
-	FS_Printf( f,"%s", "unbindall\n" );
-
-	for (const auto& kv: keys)
-	{
-		if ( kv.second.binding[ 0 ] && kv.second.binding[ 0 ][ 0 ] )
-		{
-			FS_Printf( f, "bind       %s %s\n", Key_KeynumToString( kv.first ),
-			           Cmd_QuoteString( kv.second.binding[ 0 ] ) );
-		}
-
-		for ( team = 1; team < MAX_TEAMS; ++team )
-		{
-			if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
-			{
-				FS_Printf( f, "teambind %d %s %s\n", team, Key_KeynumToString( kv.first ),
-				           Cmd_QuoteString( kv.second.binding[ team ] ) );
-			}
-		}
-	}
-}
-
-/*
-============
-Key_Bindlist_f
-
-============
-*/
-static void Key_Bindlist_f()
-{
-	int team;
-
-	for (auto& kv: keys)
-	{
-		bool teamSpecific = false;
-
-		for ( team = 1; team < MAX_TEAMS; ++team )
-		{
-			if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
-			{
-				teamSpecific = true;
-				break;
-			}
-		}
-
-		if ( !teamSpecific )
-		{
-			if ( kv.second.binding[ 0 ] && kv.second.binding[ 0 ][ 0 ] )
-			{
-				Log::Notice( "%s = %s\n", Key_KeynumToString( kv.first ), kv.second.binding[ 0 ] );
-			}
-		}
-		else
-		{
-			for ( team = 0; team < MAX_TEAMS; ++team )
-			{
-				if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
-				{
-					Log::Notice( "%s[%s] = %s\n", Key_KeynumToString( kv.first ), teamName[ team ], kv.second.binding[ team ] );
-				}
-			}
-		}
-	}
-}
-
 /*
 ============
 Key_SetKeyData
@@ -884,356 +443,6 @@ unsigned int Key_GetKeyTime()
 	return plusCommand.valid ? plusCommand.time : 0;
 }
 
-/*
-===============
-FindMatches
-
-===============
-*/
-//TODO (kangz) rework the bind commands and their completion
-static void FindMatches( const char *s )
-{
-    Cmd_OnCompleteMatch(s);
-}
-
-static void Field_TeamnameCompletion( void ( *callback )( const char *s ), int flags )
-{
-	if ( flags & FIELD_TEAM_SPECTATORS )
-	{
-		callback( "spectators" );
-	}
-
-	if ( flags & FIELD_TEAM_DEFAULT )
-	{
-		callback( "default" );
-	}
-
-	callback( "humans" );
-	callback( "aliens" );
-}
-
-/*
-===============
-Field_CompleteKeyname
-===============
-*/
-static void Field_CompleteKeyname( int flags )
-{
-	if ( flags & FIELD_TEAM )
-	{
-		Field_TeamnameCompletion( FindMatches, flags );
-	}
-
-	Key_KeynameCompletion( FindMatches );
-}
-
-/*
-===============
-Field_CompleteTeamname
-===============
-*/
-static void Field_CompleteTeamname( int flags )
-{
-	Field_TeamnameCompletion( FindMatches, flags );
-}
-
-
-
-/*
-====================
-Key_CompleteUnbind
-====================
-*/
-static void Key_CompleteUnbind( char *args, int argNum )
-{
-	if ( argNum < 4 )
-	{
-		// Skip "unbind "
-		char *p = Com_SkipTokens( args, 1, " " );
-
-		if ( p > args )
-		{
-			Field_CompleteKeyname( argNum > 2 ? 0 : FIELD_TEAM | FIELD_TEAM_SPECTATORS | FIELD_TEAM_DEFAULT );
-		}
-	}
-}
-
-/*
-====================
-Key_CompleteBind
-Key_CompleteTeambind
-====================
-*/
-static void Key_CompleteBind_Internal( char *args, int argNum, int nameArg )
-{
-	// assumption: nameArg is 2 or 3
-	char *p;
-
-	if ( argNum == nameArg )
-	{
-		// Skip "bind "
-		p = Com_SkipTokens( args, nameArg - 1, " " );
-
-		if ( p > args )
-		{
-			Field_CompleteKeyname( 0 );
-		}
-	}
-	else if ( argNum > nameArg )
-	{
-		Cmd::Args arg(args);
-		Cmd::CompletionResult res = Cmd::CompleteArgument(Cmd::Args(arg.EscapedArgs(nameArg)), argNum - nameArg - 1);
-		for (const auto& candidate : res)
-		{
-			FindMatches(candidate.first.c_str());
-		}
-	}
-}
-
-static void Key_CompleteBind( char *args, int argNum )
-{
-	Key_CompleteBind_Internal( args, argNum, 2 );
-}
-
-static void Key_CompleteTeambind( char *args, int argNum )
-{
-	if ( argNum == 2 )
-	{
-		Field_CompleteTeamname( FIELD_TEAM_SPECTATORS | FIELD_TEAM_DEFAULT );
-	}
-	else
-	{
-		Key_CompleteBind_Internal( args, argNum, 3 );
-	}
-}
-
-static void Key_CompleteEditbind( char *, int argNum )
-{
-	if ( argNum < 4 )
-	{
-		Field_CompleteKeyname( argNum > 2 ? 0 : FIELD_TEAM | FIELD_TEAM_SPECTATORS | FIELD_TEAM_DEFAULT );
-	}
-}
-
-/*
-===============
-Helper functions for Cmd_If_f & Cmd_ModCase_f
-===============
-*/
-static const struct
-{
-	char name[ 8 ];
-	unsigned short count;
-	unsigned short bit;
-	keyNum_t index;
-} modifierKeys[] =
-{
-	{ "shift", 5, 1, K_SHIFT },
-	{ "ctrl", 4, 2, K_CTRL },
-	{ "alt", 3, 4, K_ALT },
-	{ "command", 7, 8, K_COMMAND },
-	{ "cmd", 3, 8, K_COMMAND },
-	{ "mode", 4, 16, K_MODE },
-	{ "super", 5, 32, K_SUPER },
-	{ "compose", 6, 64, K_COMPOSE },
-	{ "menu", 7, 128, K_MENU },
-	{}
-};
-// Following is no. of bits required for modifiers in the above list
-// (it doesn't reflect the array length)
-static const int NUM_RECOGNISED_MODIFIERS = 8;
-
-struct modifierMask_t
-{
-	uint16_t down, up;
-	int bits;
-};
-
-static modifierMask_t getModifierMask( const char *mods )
-{
-	int i;
-	modifierMask_t mask;
-	const char *ptr;
-	static const modifierMask_t none = {0, 0, 0};
-
-	mask = none;
-
-	--mods;
-
-	while ( *++mods == ' ' ) { /* skip leading spaces */; }
-
-	ptr = mods;
-
-	while ( *ptr )
-	{
-		int invert = ( *ptr == '!' );
-
-		if ( invert )
-		{
-			++ptr;
-		}
-
-		for ( i = 0; modifierKeys[ i ].bit; ++i )
-		{
-			// is it this modifier?
-			if ( !Q_strnicmp( ptr, modifierKeys[ i ].name, modifierKeys[ i ].count )
-					&& ( ptr[ modifierKeys[ i ].count ] == ' ' ||
-						ptr[ modifierKeys[ i ].count ] == ',' ||
-						ptr[ modifierKeys[ i ].count ] == 0 ) )
-			{
-				if ( invert )
-				{
-					mask.up |= modifierKeys[ i ].bit;
-				}
-				else
-				{
-					mask.down |= modifierKeys[ i ].bit;
-				}
-
-				if ( ( mask.down & mask.up ) & modifierKeys[ i ].bit )
-				{
-					Log::Notice( "can't have %s both pressed and not pressed\n", modifierKeys[ i ].name );
-					return none;
-				}
-
-				// right, parsed a word - skip it, maybe a comma, and any spaces
-				ptr += modifierKeys[ i ].count - 1;
-
-				while ( *++ptr == ' ' ) { /**/; }
-
-				if ( *ptr == ',' )
-				{
-					while ( *++ptr == ' ' ) { /**/; }
-				}
-
-				// ready to parse the next one
-				break;
-			}
-		}
-
-		if ( !modifierKeys[ i ].bit )
-		{
-			Log::Notice( "unknown modifier key name in \"%s\"\n", mods );
-			return none;
-		}
-	}
-
-	for ( i = 0; i < NUM_RECOGNISED_MODIFIERS; ++i )
-	{
-		if ( mask.up & ( 1 << i ) )
-		{
-			++mask.bits;
-		}
-
-		if ( mask.down & ( 1 << i ) )
-		{
-			++mask.bits;
-		}
-	}
-
-	return mask;
-}
-
-static int checkKeysDown( modifierMask_t mask )
-{
-	int i;
-
-	for ( i = 0; modifierKeys[ i ].bit; ++i )
-	{
-		if ( ( mask.down & modifierKeys[ i ].bit ) && keys[ modifierKeys[ i ].index ].down == 0 )
-		{
-			return 0; // should be pressed, isn't pressed
-		}
-
-		if ( ( mask.up & modifierKeys[ i ].bit ) && keys[ modifierKeys[ i ].index ].down )
-		{
-			return 0; // should not be pressed, is pressed
-		}
-	}
-
-	return 1; // all (not) pressed as requested
-}
-
-/*
-===============
-Key_ModCase_f
-
-Takes a sequence of modifier/command pairs
-Executes the command for the first matching modifier set
-
-===============
-*/
-static void Key_ModCase_f()
-{
-	int argc = Cmd_Argc();
-	int index = 0;
-	int max = 0;
-	int count = ( argc - 1 ) / 2; // round down :-)
-	const char *v;
-
-	int mods[ 1 << NUM_RECOGNISED_MODIFIERS ];
-	// want 'modifierMask_t mods[argc / 2 - 1];' (variable array, C99)
-	// but MSVC apparently doesn't like that
-
-	if ( argc < 3 )
-	{
-		Cmd_PrintUsage( "<modifiers> <command> [<modifiers> <command>] … [<command>]", nullptr );
-		return;
-	}
-
-	while ( index < count )
-	{
-		modifierMask_t mask = getModifierMask( Cmd_Argv( 2 * index + 1 ) );
-
-		if ( mask.bits == 0 )
-		{
-			return; // parse failure (reported) - abort
-		}
-
-		mods[ index ] = checkKeysDown( mask ) ? mask.bits : 0;
-
-		if ( max < mods[ index ] )
-		{
-			max = mods[ index ];
-		}
-
-		++index;
-	}
-
-	// If we have a tail command, use it as default
-	v = ( argc & 1 ) ? nullptr : Cmd_Argv( argc - 1 );
-
-	// Search for a suitable command to execute.
-	// Search is done as if the commands are sorted by modifier count
-	// (descending) then parameter index no. (ascending).
-	for ( ; max > 0; --max )
-	{
-		int i;
-
-		for ( i = 0; i < index; ++i )
-		{
-			if ( mods[ i ] == max )
-			{
-				v = Cmd_Argv( 2 * i + 2 );
-				goto found;
-			}
-		}
-	}
-
-found:
-
-	if ( v && *v )
-	{
-		if ( *v == '/' || *v == '\\' )
-		{
-			Cmd::BufferCommandTextAfter(va("%s\n", v + 1), true);
-		}
-		else
-		{
-			Cmd::BufferCommandTextAfter(va("vstr %s\n", v), true);
-		}
-	}
-}
 
 /*
 ===================
@@ -1243,36 +452,9 @@ CL_InitKeyCommands
 void CL_InitKeyCommands()
 {
 	// register our functions
-	Cmd_AddCommand( "bind", Key_Bind_f );
-	Cmd_SetCommandCompletionFunc( "bind", Key_CompleteBind );
-	Cmd_AddCommand( "teambind", Key_Bind_f );
-	Cmd_SetCommandCompletionFunc( "teambind", Key_CompleteTeambind );
-	Cmd_AddCommand( "unbind", Key_Unbind_f );
-	Cmd_SetCommandCompletionFunc( "unbind", Key_CompleteUnbind );
-	Cmd_AddCommand( "unbindall", Key_Unbindall_f );
-	Cmd_AddCommand( "bindlist", Key_Bindlist_f );
-	Cmd_AddCommand( "editbind", Key_EditBind_f );
-	Cmd_AddCommand( "modcase", Key_ModCase_f );
-	Cmd_SetCommandCompletionFunc( "editbind", Key_CompleteEditbind );
 	Cmd_AddCommand( "setkeydata", Key_SetKeyData_f );
 }
 
-void CL_ClearKeyBinding()
-{
-	int team;
-
-	for ( team = 0; team < MAX_TEAMS; team++ )
-	{
-		for ( auto& kv: keys )
-		{
-			if ( kv.second.binding[ team ] )
-			{
-				Z_Free( kv.second.binding[ team ] );
-				kv.second.binding[ team ] = nullptr;
-			}
-		}
-	}
-}
 
 /*
 ===================
@@ -1553,27 +735,7 @@ void Key_ClearStates()
 	Key_SetCatcher( oldKeyCatcher );
 }
 
-/*
-===================
-Key_SetTeam
-===================
-*/
-void Key_SetTeam( int newTeam )
-{
-	if ( newTeam < 0 || newTeam >= MAX_TEAMS )
-	{
-		newTeam = DEFAULT_BINDING;
-	}
 
-	if ( bindTeam != newTeam )
-	{
-		Log::Debug( "%sSetting binding team index to %d",
-			Color::ToString( Color::Green ),
-			newTeam );
-	}
-
-	bindTeam = newTeam;
-}
 
 bool AnyKeyDown()
 {
