+++ src/engine/client/key_binding.cpp
@@ -1,0 +1,790 @@
+/*
+===========================================================================
+
+Daemon GPL Source Code
+Copyright (C) 1999-2010 id Software LLC, a ZeniMax Media company.
+
+This file is part of the Daemon GPL Source Code (Daemon Source Code).
+
+Daemon Source Code is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Daemon Source Code is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Daemon Source Code.  If not, see <http://www.gnu.org/licenses/>.
+
+In addition, the Daemon Source Code is also subject to certain additional terms.
+You should have received a copy of these additional terms immediately following the
+terms and conditions of the GNU General Public License which accompanied the Daemon
+Source Code.  If not, please request a copy in writing from id Software at the address
+below.
+
+If you have questions concerning this license or the applicable additional terms, you
+may contact in writing id Software LLC, c/o ZeniMax Media Inc., Suite 120, Rockville,
+Maryland 20850 USA.
+
+===========================================================================
+*/
+
+#include "common/Common.h"
+
+#include "client.h"
+#include "framework/CommandSystem.h"
+
+
+#define CLIP(t) Math::Clamp( (t), 0, MAX_TEAMS - 1 )
+
+using Keyboard::Key;
+
+
+/*
+===============
+Helper functions for Cmd_If_f & Cmd_ModCase_f
+===============
+*/
+static const struct
+{
+	char name[ 8 ];
+	unsigned short count;
+	unsigned short bit;
+	keyNum_t index;
+} modifierKeys[] =
+{
+	{ "shift", 5, 1, K_SHIFT },
+	{ "ctrl", 4, 2, K_CTRL },
+	{ "alt", 3, 4, K_ALT },
+	{ "command", 7, 8, K_COMMAND },
+	{ "cmd", 3, 8, K_COMMAND },
+	{ "mode", 4, 16, K_MODE },
+	{ "super", 5, 32, K_SUPER },
+	{ "compose", 6, 64, K_COMPOSE },
+	{ "menu", 7, 128, K_MENU },
+	{ "" }
+};
+// Following is no. of bits required for modifiers in the above list
+// (it doesn't reflect the array length)
+static const int NUM_RECOGNISED_MODIFIERS = 8;
+
+struct modifierMask_t
+{
+	uint16_t down, up;
+	int bits;
+};
+
+static modifierMask_t getModifierMask( const char *mods )
+{
+	int i;
+	modifierMask_t mask;
+	const char *ptr;
+	static const modifierMask_t none = {0, 0, 0};
+
+	mask = none;
+
+	--mods;
+
+	while ( *++mods == ' ' ) { /* skip leading spaces */; }
+
+	ptr = mods;
+
+	while ( *ptr )
+	{
+		int invert = ( *ptr == '!' );
+
+		if ( invert )
+		{
+			++ptr;
+		}
+
+		for ( i = 0; modifierKeys[ i ].bit; ++i )
+		{
+			// is it this modifier?
+			if ( !Q_strnicmp( ptr, modifierKeys[ i ].name, modifierKeys[ i ].count )
+					&& ( ptr[ modifierKeys[ i ].count ] == ' ' ||
+						ptr[ modifierKeys[ i ].count ] == ',' ||
+						ptr[ modifierKeys[ i ].count ] == 0 ) )
+			{
+				if ( invert )
+				{
+					mask.up |= modifierKeys[ i ].bit;
+				}
+				else
+				{
+					mask.down |= modifierKeys[ i ].bit;
+				}
+
+				if ( ( mask.down & mask.up ) & modifierKeys[ i ].bit )
+				{
+					Log::Notice( "can't have %s both pressed and not pressed", modifierKeys[ i ].name );
+					return none;
+				}
+
+				// right, parsed a word - skip it, maybe a comma, and any spaces
+				ptr += modifierKeys[ i ].count - 1;
+
+				while ( *++ptr == ' ' ) { /**/; }
+
+				if ( *ptr == ',' )
+				{
+					while ( *++ptr == ' ' ) { /**/; }
+				}
+
+				// ready to parse the next one
+				break;
+			}
+		}
+
+		if ( !modifierKeys[ i ].bit )
+		{
+			Log::Notice( "unknown modifier key name in \"%s\"", mods );
+			return none;
+		}
+	}
+
+	for ( i = 0; i < NUM_RECOGNISED_MODIFIERS; ++i )
+	{
+		if ( mask.up & ( 1 << i ) )
+		{
+			++mask.bits;
+		}
+
+		if ( mask.down & ( 1 << i ) )
+		{
+			++mask.bits;
+		}
+	}
+
+	return mask;
+}
+
+static int checkKeysDown( modifierMask_t mask )
+{
+	int i;
+
+	for ( i = 0; modifierKeys[ i ].bit; ++i )
+	{
+		if ( ( mask.down & modifierKeys[ i ].bit ) && keys[ modifierKeys[ i ].index ].down == 0 )
+		{
+			return 0; // should be pressed, isn't pressed
+		}
+
+		if ( ( mask.up & modifierKeys[ i ].bit ) && keys[ modifierKeys[ i ].index ].down )
+		{
+			return 0; // should not be pressed, is pressed
+		}
+	}
+
+	return 1; // all (not) pressed as requested
+}
+
+
+
+void CL_ClearKeyBinding()
+{
+	int team;
+
+	for ( team = 0; team < MAX_TEAMS; team++ )
+	{
+		for ( auto& kv: keys )
+		{
+			if ( kv.second.binding[ team ] )
+			{
+				Z_Free( kv.second.binding[ team ] );
+				kv.second.binding[ team ] = nullptr;
+			}
+		}
+	}
+}
+
+/*
+===================
+Key_SetTeam
+===================
+*/
+void Key_SetTeam( int newTeam )
+{
+	if ( newTeam < 0 || newTeam >= MAX_TEAMS )
+	{
+		newTeam = DEFAULT_BINDING;
+	}
+
+	if ( bindTeam != newTeam )
+	{
+		Log::Debug( "%sSetting binding team index to %d",
+			Color::ToString( Color::Green ),
+			newTeam );
+	}
+
+	bindTeam = newTeam;
+}
+
+/*
+===================
+Key_GetTeam
+Assumes 'three' teams: spectators, aliens, humans
+===================
+*/
+static const char *const teamName[] = { "default", "aliens", "humans", "others" };
+
+int Key_GetTeam( const char *arg, const char *cmd )
+{
+	static const struct {
+		char team;
+		char label[11];
+	} labels[] = {
+		{ 0, "spectators" },
+		{ 0, "default" },
+		{ 1, "aliens" },
+		{ 2, "humans" }
+	};
+	int t, l;
+
+	if ( !*arg ) // empty string
+	{
+		goto fail;
+	}
+
+	for ( t = 0; arg[ t ]; ++t )
+	{
+		if ( !Str::cisdigit( arg[ t ] ) )
+		{
+			break;
+		}
+	}
+
+	if ( !arg[ t ] )
+	{
+		t = atoi( arg );
+
+		if ( t != CLIP( t ) )
+		{
+			return -1;
+		}
+
+		return t;
+	}
+
+	l = strlen( arg );
+
+	for ( unsigned t = 0; t < ARRAY_LEN( labels ); ++t )
+	{
+		// matching initial substring
+		if ( !Q_strnicmp( arg, labels[ t ].label, l ) )
+		{
+			return labels[ t ].team;
+		}
+	}
+
+fail:
+	return -1;
+}
+
+/*
+===================
+Key_SetBinding
+
+team == -1 clears all bindings for the key, then sets the spec/global binding
+===================
+*/
+void Key_SetBinding( Key key, int team, const char *binding )
+{
+	char *lcbinding; // fretn - make a copy of our binding lowercase
+	// so name toggle scripts work again: bind x name BzZIfretn?
+	// resulted into bzzifretn?
+
+	if ( !key.IsBindable() )
+	{
+		return;
+	}
+
+	// free old bindings
+	if ( team == -1 )
+	{
+		// just the team-specific ones here
+		for ( team = MAX_TEAMS - 1; team; --team )
+		{
+			if ( keys[ key ].binding[ team ] )
+			{
+				Z_Free( keys[ key ].binding[ team ] );
+				keys[ key ].binding[ team ] = nullptr;
+			}
+		}
+		// team == 0...
+	}
+
+	team = CLIP( team );
+
+	if ( keys[ key ].binding[ team ] )
+	{
+		Z_Free( keys[ key ].binding[ team ] );
+	}
+
+	// set the new binding, if not null/empty
+	if ( binding && binding[ 0 ] )
+	{
+		// allocate memory for new binding
+		keys[ key ].binding[ team ] = CopyString( binding );
+		lcbinding = CopyString( binding );
+		Q_strlwr( lcbinding );  // saves doing it on all the generateHashValues in Key_GetBindingByString
+		Z_Free( lcbinding );
+	}
+	else
+	{
+		keys[ key ].binding[ team ] = nullptr;
+	}
+
+	bindingsModified = true;
+}
+
+/*
+===================
+Key_GetBinding
+
+-ve team no. = don't return the default binding
+===================
+*/
+const char *Key_GetBinding( Key key, int team )
+{
+	const char *bind;
+
+	if ( !key.IsBindable() )
+	{
+		return nullptr;
+	}
+
+	if ( team <= 0 )
+	{
+		return keys[ key ].binding[ CLIP( -team ) ];
+	}
+
+	bind = keys[ key ].binding[ CLIP( team ) ];
+	return bind ? bind : keys[ key ].binding[ 0 ];
+}
+
+
+/*
+============
+Key_WriteBindings
+
+Writes lines containing "bind key value"
+============
+*/
+void Key_WriteBindings( fileHandle_t f )
+{
+	int team;
+
+	FS_Printf( f,"%s", "unbindall\n" );
+
+	for (const auto& kv: keys)
+	{
+		if ( kv.second.binding[ 0 ] && kv.second.binding[ 0 ][ 0 ] )
+		{
+			FS_Printf( f, "bind       %s %s\n", Key_KeynumToString( kv.first ),
+					   Cmd_QuoteString( kv.second.binding[ 0 ] ) );
+		}
+
+		for ( team = 1; team < MAX_TEAMS; ++team )
+		{
+			if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
+			{
+				FS_Printf( f, "teambind %d %s %s\n", team, Key_KeynumToString( kv.first ),
+				           Cmd_QuoteString( kv.second.binding[ team ] ) );
+			}
+		}
+	}
+}
+
+namespace { // Key binding commands
+
+Cmd::CompletionResult CompleteEmbeddedCommand(const Cmd::Args& args, int startIndex, int completeIndex) {
+	// TODO: make not O(n^2)?
+	Cmd::Args trailingArgs(std::vector<std::string>(args.begin() + startIndex, args.end()));
+	return Cmd::CompleteArgument(trailingArgs, completeIndex - startIndex);
+}
+
+void CompleteTeamName(Str::StringRef prefix, Cmd::CompletionResult& completions)
+{
+	Cmd::AddToCompletion(completions, prefix, {{"spectators", ""}, {"default", ""}, {"humans", ""}, {"aliens", ""}});
+}
+
+
+void CompleteTeamAndKey(int argNum, const Cmd::Args& args, Str::StringRef prefix, Cmd::CompletionResult& res) {
+	bool hasTeam = argNum > 1 && Key_GetTeam(args.Argv(1).c_str(), "") >= 0;
+	if (argNum == 1) {
+		CompleteTeamName(prefix, res);
+	}
+	if (argNum == 1 + +hasTeam) {
+		Key_KeynameCompletion(res, prefix);
+	}
+}
+
+std::string InvalidKeyMessage(Str::StringRef name) {
+	return Str::Format("\"%s\" is not a recognized key", name);
+}
+std::string InvalidTeamMessage(Str::StringRef name) {
+	return Str::Format("\"%s\" is not a valid team", name);
+}
+
+class BindListCmd: public Cmd::StaticCmd
+{
+public:
+	BindListCmd():
+		StaticCmd("bindlist", Cmd::KEY_BINDING, "Lists all key bindings")
+	{}
+
+	void Run(const Cmd::Args&) const OVERRIDE
+	{
+		for (auto& kv: keys)
+		{
+			bool teamSpecific = false;
+
+			for ( int team = 1; team < MAX_TEAMS; ++team )
+			{
+				if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
+				{
+					teamSpecific = true;
+					break;
+				}
+			}
+
+			if ( !teamSpecific )
+			{
+				if ( kv.second.binding[ 0 ] && kv.second.binding[ 0 ][ 0 ] )
+				{
+					Log::Notice( "%s = %s", Key_KeynumToString( kv.first ), kv.second.binding[ 0 ] );
+				}
+			}
+			else
+			{
+				for ( int team = 0; team < MAX_TEAMS; ++team )
+				{
+					if ( kv.second.binding[ team ] && kv.second.binding[ team ][ 0 ] )
+					{
+						Log::Notice( "%s[%s] = %s", Key_KeynumToString( kv.first ), teamName[ team ], kv.second.binding[ team ] );
+					}
+				}
+			}
+		}
+	}
+};
+
+class BindCmd: public Cmd::StaticCmd
+{
+	bool teambind;
+
+	void ShowBind(Key b, std::function<bool(int)> teamFilter) const
+	{
+		bool bound = false;
+
+		for ( int i = 0; i < MAX_TEAMS; ++i )
+		{
+			if ( teamFilter(i) && keys[ b ].binding[ i ] )
+			{
+				Print( "\"%s\"[%s] = %s", Key_KeynumToString(b), teamName[ i ],
+					   Cmd_QuoteString( keys[ b ].binding[ i ] ) );
+				bound = true;
+			}
+		}
+
+		if ( !bound )
+		{
+			Print( "\"%s\" is not bound", Key_KeynumToString(b) );
+		}
+	}
+
+public:
+	BindCmd(Str::StringRef name, Str::StringRef desc, bool teambind):
+		StaticCmd(name, Cmd::KEY_BINDING, desc), teambind(teambind)
+	{}
+
+	void Run(const Cmd::Args& args) const OVERRIDE
+	{
+		int		c;
+		const char *key;
+		int		team = -1;
+
+		c = args.Argc();
+
+		if (c < 2 + +teambind) {
+			PrintUsage(args, teambind ? "<team> <key> [<command>]" : "<key> [<command>]", "attach a command to a key");
+			return;
+		}
+
+		if ( teambind )
+		{
+			team = Key_GetTeam( args.Argv( 1 ).c_str(), args.Argv( 0 ).c_str() );
+			if (team < 0) {
+				Print(InvalidTeamMessage(args.Argv(1)));
+				return;
+			}
+		}
+
+		key = args.Argv( 1 + +teambind ).c_str();
+		Key b = Key_StringToKeynum( key );
+
+		if ( !b.IsBindable() )
+		{
+			Print(InvalidKeyMessage(key));
+			return;
+		}
+
+		if (c == 2 + +teambind) {
+			if (teambind) {
+				ShowBind(b, [team](int t) { return t == team; });
+			} else {
+				ShowBind(b, [](int) { return true; });
+			}
+			return;
+		}
+
+		Key_SetBinding( b, team, args.ConcatArgs( 2 + +teambind ).c_str() );
+	}
+
+	Cmd::CompletionResult Complete(int argNum, const Cmd::Args& args, Str::StringRef prefix) const OVERRIDE {
+		Cmd::CompletionResult res;
+		if (teambind && argNum == 1) {
+			CompleteTeamName(prefix, res);
+		} else if (argNum == 1 + +teambind) {
+			Key_KeynameCompletion(res, prefix);
+		} else {
+			DAEMON_ASSERT_GE(argNum, 2 + +teambind);
+			res = CompleteEmbeddedCommand(args, 2 + +teambind, argNum);
+		}
+		return res;
+	}
+};
+
+
+class EditBindCmd: public Cmd::StaticCmd
+{
+public:
+	EditBindCmd():
+		StaticCmd("editbind", Cmd::KEY_BINDING, "Puts a key binding in the text field for editing")
+	{}
+
+	void Run(const Cmd::Args& args) const OVERRIDE
+	{
+		std::u32string buf;
+		int			b;
+		int			team = -1;
+
+		b = args.Argc();
+
+		if ( b < 2 || b > 3 )
+		{
+			PrintUsage(args, "[<team>] <key>");
+			return;
+		}
+
+		if ( b > 2 )
+		{
+			team = Key_GetTeam( args.Argv( 1 ).c_str(), "editbind" );
+
+			if ( team < 0 )
+			{
+				Print(InvalidTeamMessage(args.Argv(1)));
+				return;
+			}
+		}
+
+		const char *key = args.Argv( b - 1 ).c_str();
+		Key k = Key_StringToKeynum( key );
+
+		if ( !k.IsBindable() )
+		{
+			Print(InvalidKeyMessage(key));
+			return;
+		}
+
+		if ( team >= 0 )
+		{
+			buf = Str::UTF8To32("/teambind ");
+			buf += Str::UTF8To32( teamName[ team ] );
+			buf += Str::UTF8To32(" ");
+		}
+		else
+		{
+			buf = Str::UTF8To32("/bind ");
+		}
+
+		buf += Str::UTF8To32( Key_KeynumToString( k ) );
+		buf += Str::UTF8To32(" ");
+
+		const char *binding = Key_GetBinding( k, -team );
+		if ( binding )
+		{
+			buf += Str::UTF8To32( Cmd::Escape( std::string( binding ) ) );
+		}
+
+		// FIXME: use text console if that's where the editbind command was entered
+		Con_OpenConsole_f();
+		g_consoleField.SetText( buf );
+	}
+
+	Cmd::CompletionResult Complete(int argNum, const Cmd::Args& args, Str::StringRef prefix) const OVERRIDE {
+		Cmd::CompletionResult res;
+		CompleteTeamAndKey(argNum, args, prefix, res);
+		return res;
+	}
+};
+
+class UnbindCmd: public Cmd::StaticCmd
+{
+public:
+	UnbindCmd():
+		StaticCmd("unbind", Cmd::KEY_BINDING, "Removes a key binding")
+	{}
+
+	void Run(const Cmd::Args& args) const OVERRIDE
+	{
+		int b = args.Argc();
+		int team = -1;
+
+		if ( b < 2 || b > 3 )
+		{
+			PrintUsage(args, "[<team>] <key>", "remove commands from a key");
+			return;
+		}
+
+		if ( b > 2 )
+		{
+			team = Key_GetTeam( args.Argv( 1 ).c_str(), "unbind" );
+
+			if ( team < 0 )
+			{
+				Print(InvalidTeamMessage(args.Argv(1)));
+				return;
+			}
+		}
+
+		Key key = Key_StringToKeynum( args.Argv( b - 1 ).c_str() );
+
+		if ( !key.IsBindable() )
+		{
+			Print(InvalidKeyMessage(args.Argv(b - 1)));
+			return;
+		}
+
+		Key_SetBinding( key, team, nullptr );
+	}
+
+	Cmd::CompletionResult Complete(int argNum, const Cmd::Args& args, Str::StringRef prefix) const OVERRIDE {
+		Cmd::CompletionResult res;
+		CompleteTeamAndKey(argNum, args, prefix, res);
+		return res;
+	}
+};
+
+class UnbindAllCmd: public Cmd::StaticCmd
+{
+public:
+	UnbindAllCmd():
+		StaticCmd("unbindall", Cmd::KEY_BINDING, "Removes all key bindings")
+	{}
+
+	void Run(const Cmd::Args&) const OVERRIDE
+	{
+		for (auto& kv : keys)
+		{
+			Key_SetBinding( kv.first, -1, nullptr );
+		}
+	}
+};
+
+class ModcaseCmd: public Cmd::StaticCmd
+{
+public:
+	ModcaseCmd():
+		StaticCmd("modcase", Cmd::KEY_BINDING, "Conditionally performs an action based on modifier keys")
+	{}
+
+	void Run(const Cmd::Args& args) const OVERRIDE
+	{
+		int argc = args.Argc();
+		int index = 0;
+		int max = 0;
+		int count = ( argc - 1 ) / 2; // round down :-)
+		const char *v;
+
+		int mods[ 1 << NUM_RECOGNISED_MODIFIERS ];
+		// want 'modifierMask_t mods[argc / 2 - 1];' (variable array, C99)
+		// but MSVC apparently doesn't like that
+
+		if ( argc < 3 )
+		{
+			PrintUsage( args, "<modifiers> <command> [<modifiers> <command>] … [<command>]" );
+			return;
+		}
+
+		while ( index < count )
+		{
+			modifierMask_t mask = getModifierMask( args.Argv( 2 * index + 1 ).c_str() );
+
+			if ( mask.bits == 0 )
+			{
+				return; // parse failure (reported) - abort
+			}
+
+			mods[ index ] = checkKeysDown( mask ) ? mask.bits : 0;
+
+			if ( max < mods[ index ] )
+			{
+				max = mods[ index ];
+			}
+
+			++index;
+		}
+
+		// If we have a tail command, use it as default
+		v = ( argc & 1 ) ? nullptr : args.Argv( argc - 1 ).c_str();
+
+		// Search for a suitable command to execute.
+		// Search is done as if the commands are sorted by modifier count
+		// (descending) then parameter index no. (ascending).
+		for ( ; max > 0; --max )
+		{
+			int i;
+
+			for ( i = 0; i < index; ++i )
+			{
+				if ( mods[ i ] == max )
+				{
+					v = args.Argv( 2 * i + 2 ).c_str();
+					goto found;
+				}
+			}
+		}
+
+	found:
+
+		if ( v && *v )
+		{
+			if ( *v == '/' || *v == '\\' )
+			{
+				Cmd::BufferCommandTextAfter(va("%s\n", v + 1), true);
+			}
+			else
+			{
+				Cmd::BufferCommandTextAfter(va("vstr %s\n", v), true);
+			}
+		}
+	}
+
+	// TODO: completion
+};
+
+const BindListCmd BindListCmdRegistration;
+const BindCmd BindCmdRegistration{"bind", "Set or view command to be executed when a key is pressed", false};
+const BindCmd TeambindCmdRegistration{"teambind", "Set key binding for a specific team", true};
+const EditBindCmd EditBindCmdRegistration;
+const UnbindCmd UnbindCmdRegistration;
+const UnbindAllCmd UnbindAllCmdRegistration;
+const ModcaseCmd ModcaseCmdRegistration;
+
+} // namespace (key binding commands)
+
